feat(TMG-3003): separate tests for utils functions
