+++ packages/utils/__tests__/check-for-symphony-experiment.test.js
@@ -1,40 +1,4 @@
-import getCookieValue from "../src/get-cookie-value";
-import getBase64CookieValue from "../src/get-base64-cookie-value";
 import checkForSymphonyExperiment from "../src/check-for-symphony-experiment";
-
-describe("getCookieValue", () => {
-  beforeEach(() => {
-    document.cookie = "testCookie=testValue";
-  });
-
-  it("returns the value of the specified cookie", () => {
-    expect(getCookieValue("testCookie")).toBe("testValue");
-  });
-
-  it("returns null if the cookie is not found", () => {
-    expect(getCookieValue("nonExistentCookie")).toBe(null);
-  });
-});
-
-describe("getBase64CookieValue", () => {
-  beforeEach(() => {
-    const base64Value = btoa(JSON.stringify({ key: "value" }));
-    document.cookie = `base64Cookie=${base64Value}`;
-  });
-
-  it("returns the parsed value of the base64 encoded cookie", () => {
-    expect(getBase64CookieValue("base64Cookie")).toEqual({ key: "value" });
-  });
-
-  it("returns undefined if the cookie is not found", () => {
-    expect(getBase64CookieValue("nonExistentCookie")).toBe(undefined);
-  });
-
-  it("returns undefined if the cookie value is not valid base64", () => {
-    document.cookie = "invalidBase64Cookie=invalidValue";
-    expect(getBase64CookieValue("invalidBase64Cookie")).toBe(undefined);
-  });
-});
 
 describe("checkForSymphonyExperiment", () => {
   it("returns true if the Symphony experiment is active", () => {
